% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unibisect.R
\name{bisection}
\alias{bisection}
\alias{unibisect}
\alias{fast_unibisect}
\alias{simple_unibisect}
\title{Univariate bisection method}
\usage{
unibisect(f, a, b, tol = 10 * .Machine$double.eps, maxit = 100L)

fast_unibisect(f, a, b, tol = 10 * .Machine$double.eps, bisections = 8L,
  maxit = 100L)

simple_unibisect(f, a, b, tol = 10 * .Machine$double.eps, maxit = 100L)
}
\arguments{
\item{f}{function}

\item{a}{lower bound}

\item{b}{upper bound}

\item{tol}{tolerance, defaults to 10*.Machine$double.eps}

\item{maxit}{maximum number of iterations}
}
\value{
a list
}
\description{
Find the root of a function between two points a and b using the
bisection method.
}
\examples{

simple_unibisect(function(x) x^2 - 2, 0, 2)
unibisect(function(x) x^2 - 2, 0, 2)
fast_unibisect(function(x) x^2 - 2, 0, 2)


f <- function(x) x^2 - 2
a <- 0; b <- 2
(out <- unibisect(f, a, b))

curve(f(x), col = "red", from = a, to = b)
with(out$evals, points(x, fx))
out$n_evals # = number of function calls
plot(1:out$n_evals, out$evals$fx)




f <- sin
a <- .1; b <- 2*pi - .2
(out <- unibisect(f, a, b))

curve(f(x), col = "red", from = a, to = b)
with(out$evals, points(x, fx))
out$n_evals # = number of function calls
plot(1:out$n_evals, out$evals$fx)




f <- function(x) (x-.24) * (x - .51) * (x - .76)
a <- 0; b <- 1
(out <- unibisect(f, a, b))

curve(f(x), col = "red", from = a, to = b)
with(out$evals, points(x, fx))
out$n_evals # = number of function calls
plot(1:out$n_evals, out$evals$fx)



f <- function(x) pbeta(x, 90, 10) - .5
a <- 0; b <- 1
(out <- fast_unibisect(f, 0, 1))

curve(f(x), col = "red", from = 0, to = 1)
with(out$evals, points(x, fx))

library(ggplot2)
ggplot(out$evals, aes(x, fx, color = method)) +
  stat_function(fun = f, color = "black") +
  geom_point()
}
