% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_unibisect.R
\name{fast_unibisect}
\alias{fast_unibisect}
\title{Accelerated univariate bisection method}
\usage{
fast_unibisect(f, a, b, tol = 10 * .Machine$double.eps, bisections = 8L,
  maxit = 100L)
}
\arguments{
\item{f}{function}

\item{a}{lower bound}

\item{b}{upper bound}

\item{tol}{tolerance, defaults to 10*.Machine$double.eps}

\item{maxit}{maximum number of iterations}
}
\value{
a list
}
\description{
Find the root of a function between two points a and b using the
bisection method followed by the secant method.
}
\examples{

unibisect(function(x) x^2 - 2, 0, 2)
fast_unibisect(function(x) x^2 - 2, 0, 2)



f <- function(x) pbeta(x, 90, 10) - .5
a <- 0; b <- 1
(out <- fast_unibisect(f, 0, 1))

curve(f(x), col = "red", from = 0, to = 1)
with(out$evals, points(x, fx))

library(ggplot2)
ggplot(out$evals, aes(x, fx, color = method)) +
  stat_function(fun = f, color = "black") +
  geom_point()


}
