% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinewton.R
\name{multinewton}
\alias{multinewton}
\title{Univariate Newton method}
\usage{
multinewton(f, df, x0, tol = 10 * .Machine$double.eps, maxit = 100L)
}
\arguments{
\item{f}{function}

\item{df}{function; Jacobian matrix of f}

\item{x0}{initial value}

\item{tol}{tolerance, defaults to 10*.Machine$double.eps}

\item{maxit}{maximum number of iterations}
}
\value{
a list
}
\description{
Find the root of a function using Newton's method.
}
\examples{


multinewton(function(x) x^2 - 2, function(x) 2*x, 2)



# intersection of a plane, hyperboloid, and cone
f <- function(v) {
  x <- v[1]; y <- v[2]
  c(x^2 + y, x*y - 2*y^2)
}
df <- function(v) {
  x <- v[1]; y <- v[2]
  matrix(c(
    2*x,       1,
      y, x - 4*y
  ), nrow = 2, byrow = TRUE)
}
x0 <- c(2, 2)
f(x0)
df(x0)

out <- multinewton(f, df, x0)
str(out, 1)



# intersection of a plane, hyperboloid, and cone
# true solutions =
#   c(-3/sqrt(2), 0,  3/sqrt(2))
#   c( 3/sqrt(2), 0, -3/sqrt(2))
# corresponding to the nonlinear system
# x + y + z = 0
# x^2 - y^2 + z^2 = 9,
# x^2 + y^2 - z^2 = 0
f <- function(v) {
  x <- v[1]; y <- v[2]; z <- v[3]
  c(x + y + z, x^2 - y^2 + z^2 - 9, x^2 + y^2 - z^2)
}
df <- function(v) {
  x <- v[1]; y <- v[2]; z <- v[3]
  matrix(c(
      1,    1,    1,
    2*x, -2*y,  2*z,
    2*x,  2*y, -2*z
  ), nrow = 3, byrow = TRUE)
}
x0 <- c(2, 2, 2)
f(x0)
df(x0)
out <- multinewton(f, df, x0)
str(out, 1)
c( 3/sqrt(2), 0, -3/sqrt(2))
out$root


}
